requirements.txt
# ================================
# CORE FRAMEWORK & WEB
# ================================
Flask>=3.0.0,<4.0.0
Flask-Cors>=4.0.0,<5.0.0
Flask-Limiter>=3.5.0,<4.0.0
Flask-Login>=0.6.0,<1.0.0
Flask-Migrate>=4.0.0,<5.0.0
Flask-SocketIO>=5.0.0,<6.0.0
Flask-SQLAlchemy>=3.0.0,<4.0.0
Flask-WTF>=1.2.0,<2.0.0
Werkzeug>=3.0.0,<4.0.0
Jinja2>=3.1.0,<4.0.0
MarkupSafe>=2.1.0,<3.0.0
WTForms>=3.2.0,<4.0.0
itsdangerous>=2.2.0,<3.0.0

# ================================
# DATABASE & ORM
# ================================
SQLAlchemy>=2.0.0,<3.0.0
alembic>=1.10.0,<2.0.0
psycopg2-binary>=2.9.0,<3.0.0
greenlet>=3.0.0,<4.0.0
Mako>=1.3.0,<2.0.0

# ================================
# AI SERVICES (Simple & Lightweight)
# ================================
# ollama>=0.3.0,<1.0.0  # Not needed - we use requests for Ollama HTTP API calls
requests>=2.30.0,<3.0.0
urllib3>=2.0.0,<3.0.0

# ================================
# NLP & TEXT PROCESSING
# ================================
spacy>=3.7.0,<4.0.0
spacy-legacy>=3.0.0,<4.0.0
spacy-loggers>=1.0.0,<2.0.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
pyinflect>=0.5.1,<1.0.0
nltk>=3.8.0,<4.0.0
textstat>=0.7.0,<1.0.0
regex>=2023.0.0,<2025.0.0

# Semantic similarity for metadata classification (CPU-only)
# NOTE: For OpenShift deployment, install PyTorch CPU-only version first:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
# Then install the rest of requirements.txt
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
sentence-transformers>=2.2.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0

# Advanced caching for performance optimization
cachetools>=5.3.0,<6.0.0

# spaCy dependencies
thinc>=8.2.0,<9.0.0
blis>=1.0.0,<2.0.0
catalogue>=2.0.0,<3.0.0
confection>=0.1.0,<1.0.0
cymem>=2.0.0,<3.0.0
murmurhash>=1.0.0,<2.0.0
preshed>=3.0.0,<4.0.0
srsly>=2.4.0,<3.0.0
typer>=0.12.0,<1.0.0
wasabi>=1.1.0,<2.0.0
weasel>=0.4.0,<1.0.0

# ================================
# FILE PROCESSING & PARSING
# ================================
PyMuPDF>=1.23.0,<2.0.0
python-docx>=1.1.0,<2.0.0
PyPDF2>=3.0.0,<4.0.0
python-magic>=0.4.25,<1.0.0
filetype>=1.2.0,<2.0.0
Pillow>=10.0.0,<11.0.0

# Markup & Documentation
Markdown>=3.5.0,<4.0.0
markdown-it-py>=3.0.0,<4.0.0
mdurl>=0.1.0,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<5.0.0
soupsieve>=2.5.0,<3.0.0

# ================================
# PDF REPORT GENERATION 
# ================================
reportlab>=4.0.0,<5.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
pandas>=2.0.0,<3.0.0

# ================================
# PRODUCTION SERVER
# ================================
gunicorn>=23.0.0,<24.0.0
gevent>=24.2.0,<25.0.0

# ================================
# WEBSOCKETS & REAL-TIME
# ================================
python-socketio>=5.8.0,<6.0.0
python-engineio>=4.7.0,<5.0.0
simple-websocket>=1.0.0,<2.0.0
bidict>=0.23.0,<1.0.0
wsproto>=1.2.0,<2.0.0

# ================================
# HTTP & NETWORKING
# ================================
httpx>=0.25.0,<1.0.0
httpcore>=1.0.0,<2.0.0
h11>=0.14.0,<1.0.0
certifi>=2023.0.0,<2026.0.0
charset-normalizer>=3.2.0,<4.0.0
idna>=3.4.0,<4.0.0
sniffio>=1.3.0,<2.0.0
anyio>=4.0.0,<5.0.0

# ================================
# UTILITIES & CORE DEPENDENCIES
# ================================
click>=8.1.0,<9.0.0
PyYAML>=6.0.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0
python-dateutil>=2.8.0,<3.0.0
pytz>=2023.3,<2026.0
packaging>=23.0.0,<25.0.0
setuptools>=68.0.0,<81.0.0

# Configuration & Validation
pydantic>=2.4.0,<3.0.0
pydantic_core>=2.10.0,<3.0.0
annotated-types>=0.6.0,<1.0.0

# System & File Utilities
psutil>=5.9.0,<6.0.0
blinker>=1.6.0,<2.0.0
six>=1.16.0,<2.0.0

# Type System
typing_extensions>=4.8.0,<5.0.0

# Progress Display
tqdm>=4.66.0,<5.0.0

# ================================
# TESTING (Optional)
# ================================
pytest>=7.4.0,<9.0.0
pytest-flask>=1.2.0,<2.0.0